Separate concerns for the database
  - move connection to the database into database/index.js
  - move schema definition to database/model.js
  - update script to seed database
