+++ package.json
@@ -23,7 +23,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "server": "nodemon server/index.js",
-    "seed": "node database/index.js",
+    "seed": "node database/seed.js",
     "react-dev": "webpack --watch"
   },
   "keywords": [],
